fixed configuration doesn't exist error and stack overflow
